% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_gpd.R
\name{fit_gpd}
\alias{fit_gpd}
\title{Fit Generalized Pareto Distribution (GPD) with Summary Stats}
\usage{
fit_gpd(data)
}
\arguments{
\item{data}{Numeric vector of data points.}
}
\value{
A list containing the fitted distribution parameters and summary statistics.
}
\description{
Fits a Generalized Pareto Distribution (GPD) to given data and provides summary statistics.
}
\examples{
data <- rgamma(100, shape = 2, scale = 1)
fit_gpd(data)
}
